requirements.txt
batchgenerators==0.25
cachetools==5.4.0
charset-normalizer==3.3.2
clearml==1.16.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.1.1
cycler==0.12.1
Deprecated==1.2.14
dicom2nifti==2.4.11
docker==7.1.0
einops==0.8.0
entrypoints==0.4
filelock==3.15.4
fire==0.6.0
fonttools==4.53.1
fsspec==2024.6.1
furl==2.1.3
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.65.1
gunicorn==22.0.0
huggingface-hub==0.24.2
idna==3.7
imageio==2.34.2
importlib_metadata==7.2.1
importlib_resources==6.4.0
itk==5.4.0
itk-core==5.4.0
itk-filtering==5.4.0
itk-io==5.4.0
itk-numerics==5.4.0
itk-registration==5.4.0
itk-segmentation==5.4.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
linecache2==1.0.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
MedPy==0.5.2
mlflow==2.14.3
monai==1.3.2
mpmath==1.3.0
networkx==3.1
nibabel==5.2.1
nnunet==1.7.1
nptyping==2.5.0
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
orderedmultidict==1.0.1
packaging==24.1
pandas==2.0.3
pathlib2==2.3.7.post1
pillow==10.4.0
pkgutil_resolve_name==1.3.10
protobuf==4.25.4
psutil==6.0.0
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydicom==2.4.4
PyJWT==2.8.0
pynrrd==1.0.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-gdcm==3.0.24.1
pytorch-ignite==0.5.0.post2
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
querystring-parser==1.2.4
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.19.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
SimpleITK==2.3.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.1
sympy==1.13.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.4
traceback2==1.4.0
transformers==4.43.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
unittest2==1.1.0
urllib3==2.2.2
wrapt==1.16.0
zipp==3.19.2
